app/requirements.txt
Flask
pytest==3.0.5
pytest-cov==2.4.0
mysql-connector
nano